Docker configuration for development
This provides a docker-compose setup with the following images:

- source, which does a live map of the project directory to
  `/var/vhosts/framework.zend.com`.
- fpm, using php:7.0-fpm as the base, and consuming the `source` image so as to
  use the source code.
- nginx, along with essentially our production nginx configuration
  (minus HTTP -> HTTPS redirection); it maps its own port 80 to the host's port
  8080, and also consumes the `source` image in order to allow logging and to
  provide a document root to the vhost. nginx vhost configuration is provided in
  `.docker/nginx/default.conf`, and copied into `/etc/nginx/conf.d/`.

In order to make this work effectively, I removed the rewrite rule that stripped
trailing slashes, and moved it into a new middleware,
`StripTrailingSlashMiddleware`. I may do this with the manual rewrite as well,
and/or make this development configuration (so we can rely on the rewrites in
production, and have something workable in development).

The README now covers usage of docker-compose.