Use ConfigAggregator and zf-development-mode
Updated site to use ConfigAggregator-based configuration, instead of
zend-stdlib globbing/merging. Also added support for
zf-development-mode.

One discovery was that the way configuration for the manual was
occuring resulted in blockage of zf-development-mode when specific keys
were missing. I resolved that by building a trait for aggregating manual
paths, versions, etc., and this is now used by the ManualFactory and
SwitchVersionFactory to build the necessary configuration; this also
means that this configuration does not need to be built for every
request, but also that it's no longer cached.