Created post on zend-permissions-acl
Wanted to use table formatting, so added
webuni/commonmark-table-extension, and updated all places where
FrontYAML is used to add this extension.

Because it creates a plain "<table>" tag, I also updated the
AbstractCollection::getFromFile method to post-process the generated
HTML and add in relevant bootstrap classes in order to create a styled
table.