Fixed issue when formatting "updated" field in blog posts
PHP 7.0 raises a "non-well-formed numeric value" notice when calling
`date()` with an actual date string. Interestingly, the `date` field is
cast correctly, but not the `updated` field, so the change needed to be
made in the template. It now has a ternary to look for a numeric value,
and, if found, continues to use `date()`; otherwise, it passes the value
to the `DateTime` constructor, and calls `format()` on the generated
instance.