Adds general support and long-term-support policies
This patch rewrites our `/long-term-support` page to list both a
general component support policy, as well as an LTS policy based on our
skeleton applications.

Part of this is a table listing the status for each of the various
components. This requires a command for fetching the various packages we
release under the ZF and ZF-Campus aeguses, as well as all tags and
their release dates; the command processes this into a set of package
details to display.

The command is part of a new symfony/console command, `bin/zfweb.php`,
and can be invoked with `bin/zfweb.php lts:build
data/long-term-support.json` or, more simply, `composer lts:build`.