Adds command for fetching raw tag data
`zfweb.php lts:fetch-tag-data` fetches the raw tag data from GitHub, and
spits it out as a PHP file to STDOUT, allowing the user to cache the
data for later processing. This is primarily of use when testing the
lts:build command, as the data can be used with the
`CachedConfigRepository` implementation of `RepositoryInterface` in
order to provide the data to process.