Fixes issue in display of skeleton in LTS table
Not sure why this was not working on production. This change modifies
the logic to no longer use list dereferencing and instead use explode +
array_pop to get the skeleton name (vs package name).

Additionally, lack of a support date is now represented as an empty
string, not "N/A", to make sure the value stands out better.