qa: Use custom push-to-deploy
Since GitHub no longer supports integrations, and we have no easy way to
setup AWS CodePipeline or trigger CodeDeploy from Travis-CI, we will now
use a custom push-to-deploy system.

This patch rewrites the various relevant bits from the CodeDeploy setup
to work with the new push-to-deploy system. It requires adding the
following ENV variables to the repository configuration on Travis:

- DEPLOY_TOKEN
- SITE_NAME