fix: Do not `count()` the result of a `NodeList::current()` operation
A number of `DomQuery::queryXpath()` calls are made within the
`ManualAction`, which returns a `NodeList`. Many of these perform an
immediate call on its `current()` method, which in turn returns a PHP
`DOMNode`.

Unfortunately, a number of conditionals on these results were performing
a `if (count($item))`, which fails as not all `DOMNode` types are also
`Countable`. Further, in most cases, we do not care about a _count_, but
rather that we got a `DOMNode` of some sort back.

On PHP 7.2, calling `count()` on a non-countable now raises an error, so
we need to change the logic.